Add in filtering capacity
Extra one pass for mind. Should be fine? Currently will assume LD reference is well QCed and only apply the filtering on the target to reduce the complexity of the code.

If we want to add in the filtering for the LD, we will need an extra pass to ensure both LD and target has the same gene after filtering)